fix: 메모리 연산 -> db 쿼리로 결과값 추출

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -147,8 +147,8 @@ class BookServiceTest @Autowired constructor(
 
         //then
         assertThat(results).hasSize(2)
-        assertCount(results, BookType.COMPUTER, 2)
-        assertCount(results, BookType.SCIENCE, 1)
+        assertCount(results, BookType.COMPUTER, 2L)
+        assertCount(results, BookType.SCIENCE, 1L)
 
 /*        val computerDto = results.first { result -> result.type == BookType.COMPUTER }
         assertThat(computerDto.count).isEqualTo(2)
@@ -157,7 +157,7 @@ class BookServiceTest @Autowired constructor(
         assertThat(scienceDto.count).isEqualTo(1)*/
     }
 
-    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
         assertThat(results.first { result -> result.type == type }.count).isEqualTo(count)
     }
 }
